Refactor project structure, migrate to unified UI package, and update dependencies.
