+++ apps/v2/.eslintrc.js
@@ -1,3 +1,8 @@
 module.exports = {
+  root: true,
   extends: ["@repo/eslint-config/next.js"],
+  parser: "@typescript-eslint/parser",
+  parserOptions: {
+    project: true,
+  },
 };
