+++ apps/v2/package.json
@@ -10,16 +10,9 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "@radix-ui/react-dialog": "^1.1.2",
-    "@radix-ui/react-dropdown-menu": "^2.1.2",
-    "@radix-ui/react-icons": "^1.3.0",
-    "@radix-ui/react-popover": "1.1.2",
-    "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/themes": "^3.1.4",
     "@repo/ui": "*",
     "@vercel/speed-insights": "^1.0.12",
     "gray-matter": "^4.0.3",
-    "lucide-react": "^0.446.0",
     "next": "^14.2.3",
     "next-themes": "^0.3.0",
     "react": "^18.2.0",
@@ -30,7 +23,6 @@
   "devDependencies": {
     "@next/eslint-plugin-next": "^14.2.3",
     "@repo/eslint-config": "*",
-    "@repo/tailwind-config": "*",
     "@repo/typescript-config": "*",
     "@types/node": "^20.11.24",
     "@types/react": "^18.2.61",
