+++ apps/v2/postcss.config.mjs
@@ -1,0 +1,1 @@
+export { default } from "@repo/ui/postcss.config";
