+++ apps/v2/src/app/layout.tsx
@@ -1,14 +1,14 @@
-import "../../styles/globals.css";
-import "tailwindcss/tailwind.css" // Ensure Tailwind is imported if not already
+import "@repo/ui/globals.css";
+// import "tailwindcss/tailwind.css" // Ensure Tailwind is imported if not already
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
+// import { Inter } from "next/font/google";
 import { SpeedInsights } from "@vercel/speed-insights/next";
 import { Navbar } from '../components/navbar';
 
 import { ThemeProvider } from '../components/theme-provider';
 
 
-const inter = Inter({ subsets: ["latin"] });
+// const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
   title: "Liam Germain's Personal Site",
@@ -23,7 +23,7 @@ export default function RootLayout({
 
   return (
     <html lang="en" className="h-full">
-      <body className={`${inter.className} flex flex-col h-full overflow-hidden`}>
+      <body className={`flex flex-col h-full overflow-hidden`}>
 
         <ThemeProvider
           attribute="class"
