+++ apps/v2/tailwind.config.ts
@@ -1,11 +1,1 @@
-// tailwind config is required for editor support
-
-import type { Config } from "tailwindcss";
-import sharedConfig from "../../packages/config-tailwind/tailwind.config";
-
-const config: Pick<Config, "content" | "presets"> = {
-  content: ["./src/app/**/*.tsx"],
-  presets: [sharedConfig],
-};
-
-export default config;
+export * from "@repo/ui/tailwind.config";
