+++ package.json
@@ -16,7 +16,7 @@
   "engines": {
     "node": ">=18"
   },
-  "name": "with-tailwind",
+  "name": "lmg-personal-website",
   "packageManager": "npm@10.8.3",
   "workspaces": [
     "apps/*",
