--- packages/config-tailwind/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@repo/tailwind-config",
-  "version": "0.0.0",
-  "private": true,
-  "exports": {
-    ".": "./tailwind.config.ts"
-  },
-  "devDependencies": {
-    "@repo/typescript-config": "*",
-    "tailwindcss": "^3.4.1"
-  }
-}
