+++ packages/config-typescript/base.json
@@ -13,6 +13,9 @@
     "esModuleInterop": true,
     "moduleResolution": "node",
     "resolveJsonModule": true,
+    "incremental": false,
+    "declaration": true,
+    "declarationMap": true,
     "isolatedModules": true,
     "jsx": "preserve",
     "baseUrl": ".",
