+++ packages/ui/.eslintrc.js
@@ -1,3 +1,9 @@
+/** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: ["@repo/eslint-config/react.js"],
+  root: true,
+  extends: ['@repo/eslint-config/react-internal.js'],
+  parser: '@typescript-eslint/parser',
+  rules: {
+    'no-redeclare': 'off',
+  },
 };
