--- packages/ui/src/index.ts
@@ -1,3 +1,0 @@
-// packages/ui/src/index.ts
-export * from './components/button';
-export * from './components/dropdown-menu';
